+++ UIKit/Classes/UILabel.m
@@ -27,6 +27,7 @@
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include <tgmath.h>
 #import "UILabel.h"
 #import "UIColor.h"
 #import "UIFont.h"
@@ -163,7 +164,7 @@
         drawRect.size = [_text sizeWithFont:_font constrainedToSize:maxSize lineBreakMode:_lineBreakMode];
 
         // now vertically center it
-        drawRect.origin.y = roundf((bounds.size.height - drawRect.size.height) / 2.f);
+        drawRect.origin.y = round((bounds.size.height - drawRect.size.height) / 2.f);
         
         // now position it correctly for the width
         // this might be cheating somehow and not how the real thing does it...
